+++ rumqttd/src/link/remote.rs
@@ -12,6 +12,7 @@ use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::error::Elapsed;
 use tokio::{select, time};
+use tracing::debug;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
@@ -56,7 +57,7 @@ impl<P: Protocol> RemoteLink<P> {
     pub async fn new(
         config: Arc<ConnectionSettings>,
         router_tx: Sender<(ConnectionId, Event)>,
-        tenant_id: Option<String>,
+        // tenant_id: Option<String>,
         mut network: Network<P>,
     ) -> Result<RemoteLink<P>, Error> {
         // Wait for MQTT connect packet and error out if it's not received in time to prevent
@@ -90,7 +91,7 @@ impl<P: Protocol> RemoteLink<P> {
         }
 
         let (link_tx, link_rx, notification) = Link::new(
-            tenant_id,
+            // tenant_id,
             &client_id,
             router_tx,
             clean_session,
@@ -128,7 +129,7 @@ impl<P: Protocol> RemoteLink<P> {
                         buffer.len()
                     };
 
-                    debug!("{:15.15}[I] {:20} buffercount = {}", self.client_id, "packets", len);
+                    debug!(buffercount=len, "packets");
                     self.link_tx.notify().await?;
                 }
                 // Receive from router when previous when state isn't in collision